+++ packages/datadog_flags/lib/datadog_flags.dart
@@ -13,7 +13,13 @@ library;
 export 'src/datadog_flags_config.dart'
     show DatadogFlagsConfig, DatadogFlagsSite;
 export 'src/datadog_flags.dart' show DatadogFlags;
-export 'src/flags_client.dart' show DatadogFlagsClient, FlagDetails;
+export 'src/datadog_openfeature_provider.dart' show DatadogOpenFeatureProvider;
+export 'src/flags_client.dart'
+    show
+        DatadogFlagsClient,
+        DatadogFlagsClientLifecycle,
+        DatadogFlagsClientStatus,
+        FlagDetails;
 export 'src/flags_configuration.dart' show DatadogFlagsConfiguration;
 export 'src/flags_error.dart' show FlagEvaluationError;
 export 'src/flags_store.dart' show DatadogFlagsStore, FlagsData;